Splitted code into different files + fixed asserts
